requirements.txt
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
bcrypt
python-dotenv
mysql-connector-python 